pyproject.toml
[tool.poetry]
name = "peterpy"
version = "0.1.0"
description = ""
authors = ["Peter van Dijk <peter.van.dijk@bynder.com>"]
readme = "README.md"

[tool.poetry.scripts]
peterpy = "peterpy.app:main"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.5"
asyncio = "^3.4.3"
tomlkit = "^0.12.4"
configloader = "^1.0.1"
uvloop = "^0.19.0"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.29"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
mypy = "^1.9.0"
isort = "^5.13.2"
black = "^24.4.0"
bandit = "^1.7.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
